// src/app/core/models/notification-events.enum.ts
export enum NotificationEventsEnum {
  DELEGATOR_SUBSCRIPTION = 'DELEGATOR_SUBSCRIPTION',
  EOUS_FORWARD_EMAIL_TO_USER = 'EOUS_FORWARD_EMAIL_TO_USER',
  ERPTS_SUBMITTED = 'ERPTS_SUBMITTED',
  ESTATUSES_CREATED_TXN = 'ESTATUSES_CREATED_TXN',
  ESTATUSES_CREATED_RPT = 'ESTATUSES_CREATED_RPT',
  ETXNS_ADMIN_REMOVED = 'ETXNS_ADMIN_REMOVED',
  ETXNS_ADMIN_UPDATED = 'ETXNS_ADMIN_UPDATED',
  ERPTS_INQUIRY = 'ERPTS_INQUIRY',
  ERPTS_APPROVED = 'ERPTS_APPROVED',
  EREIMBURSEMENTS_COMPLETED = 'EREIMBURSEMENTS_COMPLETED',
  EADVANCE_REQUESTS_CREATED = 'EADVANCE_REQUESTS_CREATED',
  EADVANCE_REQUESTS_UPDATED = 'EADVANCE_REQUESTS_UPDATED',
  EADVANCE_REQUESTS_INQUIRY = 'EADVANCE_REQUESTS_INQUIRY',
  EADVANCE_REQUESTS_APPROVED = 'EADVANCE_REQUESTS_APPROVED',
  EADVANCES_CREATED = 'EADVANCES_CREATED',
  EADVANCE_REQUESTS_REJECTED = 'EADVANCE_REQUESTS_REJECTED',
  ERPTS_READY_TO_PROCESS = 'ERPTS_READY_TO_PROCESS',
  ETXNS_MARKED_PERSONAL = 'ETXNS_MARKED_PERSONAL',
}